Ship miniperf-record, but only if B2G_PROFILING is enabled.
Currently, user-mode call chains are available if and only if
B2G_PROFILING is set.  It would be possible to use miniperf-record
without stacks (and the executable is currently only 10 KiB, so the cost
of allowing it would be small), but the amount of useful information
available would be quite limited.